refact: renames 'admin-ticket' OpenAPI schema back to 'ticket' to reduce duplication
