+++ api/src/paths/administrative/tickets/{ticketId}/reference/index.ts
@@ -3,7 +3,7 @@ import { Operation } from 'express-openapi';
 import { SYSTEM_ROLE } from '../../../../../constants/roles';
 import { getDBConnection } from '../../../../../database/db';
 import { type CreateTicketReferenceRequest } from '../../../../../models/ticket-reference';
-import { CreateTicketReferenceRequestSchema, TicketReferenceSchema } from '../../../../../openapi/schemas/admin-ticket';
+import { CreateTicketReferenceRequestSchema, TicketReferenceSchema } from '../../../../../openapi/schemas/ticket';
 import { defaultErrorResponses } from '../../../../../openapi/schemas/http-responses';
 import { authorizeRequestHandler } from '../../../../../request-handlers/security/authorization';
 import { TicketService } from '../../../../../services/ticket-service';
