+++ api/src/paths/administrative/tickets/{ticketId}/status/index.ts
@@ -2,7 +2,7 @@ import { RequestHandler } from 'express';
 import { Operation } from 'express-openapi';
 import { SYSTEM_ROLE } from '../../../../../constants/roles';
 import { getDBConnection } from '../../../../../database/db';
-import { TicketSchema, UpdateTicketStatusRequestSchema } from '../../../../../openapi/schemas/admin-ticket';
+import { TicketSchema, UpdateTicketStatusRequestSchema } from '../../../../../openapi/schemas/ticket';
 import { defaultErrorResponses } from '../../../../../openapi/schemas/http-responses';
 import { authorizeRequestHandler } from '../../../../../request-handlers/security/authorization';
 import { TicketService } from '../../../../../services/ticket-service';
