+++ api/src/paths/tickets/{ticketId}/index.ts
@@ -1,7 +1,7 @@
 import { RequestHandler } from 'express';
 import { Operation } from 'express-openapi';
 import { getDBConnection } from '../../../database/db';
-import { TicketWithHistorySchema } from '../../../openapi/schemas/admin-ticket';
+import { TicketWithHistorySchema } from '../../../openapi/schemas/ticket';
 import { defaultErrorResponses } from '../../../openapi/schemas/http-responses';
 import { authorizeRequestHandler } from '../../../request-handlers/security/authorization';
 import { TicketService } from '../../../services/ticket-service';
